Refactors models and improves error handling
Splits relation model into separate chat and loan models for better organization
Adds null checks for user authentication and database operations
Improves type safety in notifications and handoff functions

Enhances code maintainability through model separation and stronger error handling
