+++ firebase-project/functions/src/index.ts
@@ -16,5 +16,5 @@ export {validateTool} from "./validateTool";
 export {validateRelation} from "./validateRelation";
 export {validateLoan} from "./validateLoan";
 export {validateTool_TEST} from "./validateTool_TEST";
-export {acceptHandoff, startHandoff} from "./handoff";
+export {acceptHandoff, startHandoff} from "./handoff"; // TODO add requestReturn
 export {onUserCreate} from "./authentication";
